Fix some issues - Using number as string is much easier - Using a computed get/set for calculatedResult is easier when dealing with a string that ends with a comma - Use double in backend to be able to use comma numbers

+++ CalculateRequest.cs
@@ -2,7 +2,7 @@ namespace spectra_systems;
 
 public class CalculateRequest
 {
-    public int calculatedResult { get; set; }
-    public int currentResult { get; set; }
+    public double calculatedResult { get; set; }
+    public double currentResult { get; set; }
     public string? operand { get; set; }
 }
